refactor: move render-email-html to email hooks
