+++ src/lib/emails/templates/subscription-auto-renew.tsx
@@ -1,6 +1,6 @@
 import { Text } from "@react-email/components";
 import { OrganizationSubscription } from "~/lib/organizations/types/organization-subscription";
-import { renderEmailHtml } from "./render-email-html";
+import { renderEmailHtml } from "../hooks/render-email-html";
 
 export const getSubscriptionAutoRenewEmailTemplate = (subscription: OrganizationSubscription) => {
   const periodEndAsDate = new Date(subscription.periodEndsAt * 1000);
