+++ src/lib/emails/templates/successful-project-completion.tsx
@@ -1,5 +1,5 @@
 import { Text } from "@react-email/components";
-import { renderEmailHtml } from "./render-email-html";
+import { renderEmailHtml } from "../hooks/render-email-html";
 
 export const getSuccessfulProjectCompletionEmailTemplate = () => {
   const subject = "Congrats! Your Project is Ready 🎊";
