+++ src/lib/emails/templates/welcome.tsx
@@ -1,5 +1,5 @@
 import { Text } from "@react-email/components";
-import { renderEmailHtml } from "./render-email-html";
+import { renderEmailHtml } from "../hooks/render-email-html";
 
 export const getWelcomeEmailTemplate = () => {
   const subject = "Welcome Aboard the Audioland Express! 🎉";
